add exception handler

+++ src/main/java/com/end2end/ansimnuri/note/service/NoteReplyServiceImpl.java
@@ -8,6 +8,8 @@ import com.end2end.ansimnuri.note.domain.entity.NoteReply;
 import com.end2end.ansimnuri.note.domain.repository.NoteReplyRepository;
 import com.end2end.ansimnuri.note.domain.repository.NoteRepository;
 import com.end2end.ansimnuri.note.dto.NoteReplyDTO;
+import com.end2end.ansimnuri.util.exception.UnAuthenticationException;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -27,6 +29,7 @@ public class NoteReplyServiceImpl implements NoteReplyService {
                 .toList();
     }
 
+    @Transactional
     @Override
     public void insert(NoteReplyDTO dto, String loginId) {
         Note note = noteRepository.findById(dto.getNoteId())
@@ -37,17 +40,33 @@ public class NoteReplyServiceImpl implements NoteReplyService {
         noteReplyRepository.save(NoteReply.of(dto, note, member));
     }
 
+    @Transactional
     @Override
-    public void update(NoteReplyDTO dto) {
+    public void update(NoteReplyDTO dto, String loginId) {
         NoteReply noteReply = noteReplyRepository.findById(dto.getId())
                 .orElseThrow(() -> new IllegalArgumentException("해당 id의 쪽지 댓글이 존재하지 않습니다."));
+
+        Member member = memberRepository.findByLoginId(loginId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 로그인 아이디는 존재하지 않습니다."));
+        if(!member.getId().equals(noteReply.getMember().getId())) {
+            throw new UnAuthenticationException();
+        }
+
         noteReply.update(dto);
     }
 
+    @Transactional
     @Override
-    public void deleteById(long id) {
+    public void deleteById(long id, String loginId) {
         NoteReply noteReply = noteReplyRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 id의 쪽지 댓글이 존재하지 않습니다."));
+
+        Member member = memberRepository.findByLoginId(loginId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 로그인 아이디는 존재하지 않습니다."));
+        if(!member.getId().equals(noteReply.getMember().getId())) {
+            throw new UnAuthenticationException();
+        }
+
         noteReplyRepository.delete(noteReply);
     }
 }
